Cleanup VisualRenderer Class (#210)

+++ Source/Core/Renderer/ERS_CLASS_VisualRenderer/ERS_CLASS_ViewportMenu.h
@@ -45,7 +45,7 @@ private:
     std::map<int, std::shared_ptr<ERS_STRUCT_Shader>>* Shaders_; /**<Pointer to map containing shader information*/
 
     bool* IsEditorMode_; /**<Pointer to bool setting editor mode to enabled or disablewd*/
-    float* GameStartTime_; /**<Pointer to float setting the glfw time when the game clock started*/
+    double* GameStartTime_; /**<Pointer to float setting the glfw time when the game clock started*/
 
 
 public:
@@ -55,7 +55,7 @@ public:
      * 
      * @param SystemUtils 
      */
-    ERS_CLASS_ViewportMenu(ERS_STRUCT_SystemUtils* SystemUtils, ERS_STRUCT_ProjectUtils* ProjectUtils, float* GameStartTime, bool* IsEditorMode, std::map<int, std::shared_ptr<ERS_STRUCT_Shader>>* Shaders);
+    ERS_CLASS_ViewportMenu(ERS_STRUCT_SystemUtils* SystemUtils, ERS_STRUCT_ProjectUtils* ProjectUtils, double* GameStartTime, bool* IsEditorMode, std::map<int, std::shared_ptr<ERS_STRUCT_Shader>>* Shaders);
 
     /**
      * @brief Destroy the ers class viewportMenu object
